+++ R/annotate_textp.R
@@ -19,15 +19,19 @@
 #'
 #' @examples
 #' p1 = ggplot2::ggplot(data.frame(x=1:10,y=1:10), ggplot2::aes(x,y)) + ggplot2::geom_point()
-#' p1 + annotate_textp('Text annotation\nx=1, y=0, hjust=1', x=1, y=0, hjust=1)
-#' p1 + annotate_textp('Text annotation\nx=0.1, y=0.9, hjust=0', x=0, y=1, hjust=0)
-#' p1 + annotate_textp('Text annotation\nx = 0.5, y=0.5, hjust=0.5\nbox_just=c(0.5,0.5)', x=0.5, y=0.5, hjust=0.5, box_just=c(0.5,0.5))
-#' p1 + annotate_textp('Text annotation\nx = 0.5, y=0.5, hjust=0.5\nbox_just=c(0.5,0.5)\nsize=14, alpha=0.5', x=0.5, y=0.5, hjust=0.5, box_just=c(0.5,0.5), size=14, alpha=0.5)
+#' p1 + annotate_textp(label="x=1, y=0, hjust=1", 
+#'                     x=1, y=0, hjust=1)
+#' p1 + annotate_textp(label="x=0.1, y=0.9, hjust=0", 
+#'                     x=0, y=0.9, hjust=0)
+#' p1 + annotate_textp(label="x = 0.5, y=0.5, hjust=0.5, box_just=c(0.5,0.5)", 
+#'                     x=0.5, y=0.5, hjust=0.5, box_just=c(0.5,0.5))
+#' p1 + annotate_textp(label="x = 0.5, y=0.5, hjust=0.5, box_just=c(0.5,0.5), size=6, alpha=0.5", 
+#'                     x=0.5, y=0.5, hjust=0.5, box_just=c(0.5,0.5), size=6, alpha=0.5)
 #'
 #' @export
 #'
 
-annotate_textp = function(label, x, y, facets=NULL, hjust=0, vjust=0, color='black', alpha=NA,
+annotate_textp <- function(label, x, y, facets=NULL, hjust=0, vjust=0, color='black', alpha=NA,
                           family=thm$text$family, size=thm$text$size, fontface=1, lineheight=1.0,
                           box_just=ifelse(c(x,y)<0.5,0,1), margin=grid::unit(size/2, 'pt'), thm=ggplot2::theme_get()) {
   x = scales::squish_infinite(x)